refactor(api): move acceptCertificationCenterInvitation controller method
to src/team
