+++ api/src/team/application/certification-center-invitation/certification-center-invitation.controller.js
@@ -3,6 +3,27 @@ import { usecases } from '../../domain/usecases/index.js';
 import { certificationCenterInvitationSerializer } from '../../infrastructure/serializers/jsonapi/certification-center-invitation-serializer.js';
 
 /**
+ * @callback acceptCertificationCenterInvitatio
+ * @param request
+ * @param h
+ * @returns {Promise<void>}
+ */
+const acceptCertificationCenterInvitation = async function (request, h) {
+  const certificationCenterInvitationId = request.params.id;
+  const { code, email: rawEmail } = request.deserializedPayload;
+  const localeFromCookie = request.state?.locale;
+  const email = rawEmail.trim().toLowerCase();
+
+  await usecases.acceptCertificationCenterInvitation({
+    certificationCenterInvitationId,
+    code,
+    email,
+    localeFromCookie,
+  });
+  return h.response({}).code(204);
+};
+
+/**
  * @callback sendInvitations
  * @param request
  * @param h
@@ -34,6 +55,7 @@ const findPendingInvitations = async function (request, h) {
 };
 
 export const certificationCenterInvitationController = {
+  acceptCertificationCenterInvitation,
   cancelCertificationCenterInvitation,
   findPendingInvitations,
   sendInvitations,
